very simple, very stupid proof of concept
This was written in one hacking session at a coffeeshop. It's not the
best code, for sure, but it's a starting point. The files in lib/auto
are there for now so you can do perl -Mlib=lib -MEmail::Archive .... and
also so you can run test_run, which is the only bit of tests this code
has presently.

All that said, the hard part is done: there's no longer a blank slate
for this project in my mind.